Moved smart suggestions mechanism into its own extension module. It is disabled by default and can be enabled from configuration dialog (without restarting!).
