+++ idlesporklib/AutoComplete.py
@@ -62,8 +62,7 @@ class AutoComplete:
     menudefs = [
         ('edit', [
             ("Show Completions", "<<force-open-completions>>"),
-        ]),
-        ('shell', [('Patch _suggestions', '<<patch-suggestions>>')])
+        ])
     ]
 
     popupwait = idleConf.GetOption("extensions", "AutoComplete",
@@ -105,8 +104,6 @@ class AutoComplete:
         # delayed call.
         self._delayed_completion_id = None
         self._delayed_completion_index = None
-
-        self.text.bind("<<patch-suggestions>>", self.patch_suggestions_event)
 
     def _make_autocomplete_window(self):
         return AutoCompleteWindow.AutoCompleteWindow(self.text, self.twotabstocomplete, self.entertocomplete)
@@ -396,15 +393,6 @@ class AutoComplete:
     def get_module_completion(self, line):
         """Get module completions for the line"""
         return ModuleCompletion.module_completion(line)
-
-    def patch_suggestions_event(self, event=None):
-        self.editwin.interp.register_onrestart(self.patch_suggestions)
-        self.patch_suggestions()
-
-    @boundremotefunc
-    def patch_suggestions(self):
-        from threading import Thread
-        Thread(target=ModuleCompletion.patch_suggestions).start()
 
 
 if __name__ == '__main__':
