+++ idlesporklib/ModuleCompletion.py
@@ -205,138 +205,3 @@ class ModuleCompletion(object):
                 return mods, mods
             else:
                 return [mod for mod in mods if not mod.startswith('_')], mods
-
-    @staticmethod
-    def inspect_all_objs():
-        if ModuleCompletion.allobjs is not None:
-            return
-
-        visited = set()
-        defclass = re.compile('(class|def) ([_A-z][_A-z0-9]*)[\(:]')
-        variable = re.compile('([A-z][_A-z0-9]+)\s=')
-        objs = defaultdict(dict)
-
-        rootmodules = list(sys.builtin_module_names)
-        for name in rootmodules:
-            objs[name][('module', name)] = ('builtin', 0)
-            m = __import__(name)
-            for attr in dir(m):
-                a = getattr(m, attr)
-                if isclass(a):
-                    objs[attr][('class', name)] = ('builtin', 0)
-                elif callable(a):
-                    objs[attr][('def', name)] = ('builtin', 0)
-
-        for path in sys.path:
-            if path == '':
-                path = '.'
-
-            if os.path.isdir(path):
-                for root, dirs, nondirs in os.walk(path, followlinks=True):
-                    if '-' in root[len(path) + 1:] or root in visited:
-                        dirs[:] = []
-                        continue
-
-                    visited.add(root)
-
-                    for name in nondirs:
-                        if name.endswith('.py'):
-                            filepath = os.path.join(root, name)
-
-                            if name == '__init__.py':
-                                name = root[len(path) + 1:].split('/')[-1]
-                                modulepath = '.'.join(root[len(path) + 1:].split('/')[:-1])
-                            else:
-                                name = name[:-3]
-                                modulepath = root[len(path) + 1:].replace('/', '.')
-
-                            if modulepath.endswith('.'):
-                                modulepath = modulepath[:-1]
-
-                            if ('module', modulepath) not in objs[name]:
-                                objs[name][('module', modulepath)] = (filepath, 0)
-
-                                with open(filepath, 'r') as f:
-                                    for i, line in enumerate(f):
-                                        m = defclass.match(line)
-                                        if m:
-                                            t, sym = m.groups()
-                                            objs[sym][(t, '%s.%s' % (modulepath, name))] = (filepath, i)
-                                        else:
-                                            m = variable.match(line)
-                                            if m:
-                                                objs[m.group(1)][('var', '%s.%s' % (modulepath, name))] = (filepath, i)
-
-        ModuleCompletion.allobjs = objs
-
-    @staticmethod
-    def patch_suggestions():
-        if ModuleCompletion.patched:
-            print('already patched.')
-            return
-        elif ModuleCompletion.executing_patch:
-            print('patching already taking place...')
-            return
-
-        print('preparing suggestions...')
-
-        ModuleCompletion.executing_patch = True
-        ModuleCompletion.inspect_all_objs()
-
-        old_import_suggest = Suggest.import_suggest
-
-        def new_import_suggest(name, source):
-            old_import_suggest(name, source)
-
-            cl = candidates(name, ModuleCompletion.allobjs)
-            suggestions = []
-
-            if len(cl) > 0:
-                for word in cl:
-                    for (t, modulepath), (filepath, linenum) in ModuleCompletion.allobjs[word].items():
-                        if t == 'module':
-                            if filepath not in ['builtin', '']:
-                                tag = 'M'
-                                filelink = Links.FileLink(None, 'M', filepath, linenum + 1).create()
-                            else:
-                                tag = 'BM'
-                                filelink = 'BM'
-
-                            if modulepath == '' or filepath == 'builtin':
-                                link1 = sporktools.Links.ExecCodeLink(None, "%s" % word,
-                                                                      "import %s" % word).create()
-                                suggestions.append(("(%s) import %s" % (filelink, link1),
-                                                    "(%s) import %s" % (tag, word)))
-                            else:
-                                link1 = sporktools.Links.ExecCodeLink(None, "%s" % word,
-                                                                      "from %s import %s" % (modulepath, word)).create()
-                                suggestions.append(("(%s) from %s import %s" % (filelink, modulepath, link1),
-                                                    "(%s) from %s import %s" % (tag, modulepath, word)))
-                        else:
-                            if t == 'class':
-                                tag = 'C'
-                            elif t == 'def':
-                                tag = 'F'
-                            else:  # t == 'var'
-                                tag = 'V'
-
-                            if filepath not in ['builtin', '']:
-                                filelink = Links.FileLink(None, tag, filepath, linenum + 1).create()
-                            else:
-                                filelink = 'B' + tag
-                            link1 = sporktools.Links.ExecCodeLink(None, "%s" % word,
-                                                                  "from %s import %s" % (modulepath, word)).create()
-                            suggestions.append(("(%s) from %s import %s" % (filelink, modulepath, link1),
-                                                "(%s) from %s import %s" % (tag, modulepath, word)))
-
-            if suggestions:
-                Suggest.newline()
-                print("Here are some import suggestions:", file=sys.stderr)
-                for suggestion, realtxt in sorted(suggestions, key=lambda x: x[1]):
-                    Suggest.newline()
-                    print(suggestion, file=sys.stderr)
-
-        Suggest.import_suggest = new_import_suggest
-        ModuleCompletion.executing_patch = False
-        ModuleCompletion.patched = True
-        print('done preparing suggestions.')
