+++ idlesporklib/config-extensions.def
@@ -64,6 +64,9 @@ refresh-calltip=<KeyRelease-parenright> <KeyRelease-0>
 [ResourcesStatus]
 enable=False
 
+[SmartSuggest]
+enable=False
+
 [CodeContext]
 enable=True
 enable_shell=False
